Extend snyk patch expiry duration to 1 year (#39)
Snyk keeps prompting to fix this vulnerability, but it's a false positive, and Snyk's automated patches break the build and add several thousand unnecessary lines of patch notes to this file. We're better off doing it manually and leaving it in there in the hope that Snyk fix the buggy behaviour on their end.