[KED-1218] Add theme prop and icon button visibility prop (#80)
JavaScript component consumers of the application have requested the ability to override the theme, and to hide the theme button. This PR adds extra props to facilitate this and also allow users to hide future features like the upcoming export button.

This PR also removes leftover traces of the created_ts and message data properties, which are no longer used.

Finally, it also turns on text labels by default.