[KED-1730] Improve rendering performance (#186)
* Prevent layout calc from running twice on pageload

Though I had tried to avoid from being layout being called until after fontLoaded was ready, I missed that it was being called by the export-graph function, which meant that it was called twice on page load (i.e. on first load and then again when the font loads). I have changed this so that it now explicitly listens to fontLoaded, and won't return if fontLoaded is false. The same is now true for some of the more expensive node selectors.

* Optimise getNodeTextWidth selector

This selector is a relatively expensive one, as it involves some hefty DOM operations. It turns out (from benchmarking tests) that using D3's selectAll approach is faster than the approach that I was using.

* Set fontLoaded=true by default in mock state for testing