[KED-1597] Add web worker for graph layout calculation (#217)
This PR enables the use of web workers to asynchronously perform time-consuming calculations (e.g. for instance the dagre/newgraph layout calculation) in a separate CPU thread, in order to prevent it from blocking other operations (e.g. CSS transitions and other state updates).

This now means that we can display a loading spinner while the layout is being calculated, which should improve the user experience in larger graphs.

While this PR only adds a web worker for the layout calculation, it opens up the ability to potentially workerize other expensive calculations, and potentially display a loading spinner whenever any of them are in progress.