[KED-2092] Fix pipeline selector so that it shows the correct default pipeline (#266)
Hotfix for an issue in the v3.5.0 release of Kedro-Viz, where the default/initial pipeline behaviour is out of sync between the front and back end. This PR resolves this issue by adding support for the selected_pipeline prop in the front end, so that the default pipeline will use this prop, rather than the first pipeline in the list.

This PR also changes the selected_pipeline prop so that it is just a string ID, instead of an object containing the ID and name of the default pipeline. This is because the other info is redundant, so we felt it better to simplify it.