Hide graph when loading+switching to new pipeline (#296)
I've noticed that often, when loading the graph while changing the pipeline, it's not sufficiently obvious that the chart is still loading. This is because the sidebar updates before the graph does, which gives the impression that it's ready when it isn't yet. This change aims to prevent this by removing the old graph when loading, to make it obvious that the new graph is still being generated.