Transcoding Datasets on Kedro-Viz
Transcoded datasets were not represented correctly in Kedro-viz. Instead of displaying as a single node with different original and transcoded dataset type; they were displayed as two separate nodes without any link to one another.

In this ticket, transcoded datasets are represented as a single node on the flowchart and the metadata panel displays unique information about the transcoded datasets such as original type and transcoded type.