Fix CircleCI on main (#818)
This issue arises because of the new ticket to Extend Python 3.9, 3.10. The current cache version for python is now 3.10 (instead of 3.8) and this is failing circleCI on main.
Given that the cache_version variable isn't injecting any value at the moment, we are replacing it with 0 so it can fix the main build issue.

Following this quick fix -- we will fix the circleCI to extend support for 3.9, 3.10.