Fix modular pipelines breaking when collapsed.  (#1651)
In modular pipelines, a bug was identified when the pipelines were in collapsed mode. This issue was related to the incorrect handling of datasets due to a flawed logic in defining inputs and outputs. Here's the explanation:

Inputs of a Modular Pipeline:
An input can either be an external source or an internal one, provided it's not also used internally as an output.
If an internal input is simultaneously an internal output, it implies it's part of a sub-pipeline within the modular pipeline. Therefore, it should remain hidden in the collapsed view of the modular pipeline and not be displayed as an input.
Outputs of a Modular Pipeline:
An output is typically any external dataset used by other pipelines.
In the Kedro framework, an external output must be explicitly defined as an output of the modular pipeline. It doesn't have a namespace.
Internal outputs that also function as internal inputs are treated as components of the modular pipeline. These are not visible as outputs in the collapsed view but become visible when the pipeline is expanded.
This improved handling ensures a more accurate representation and functionality of modular pipelines, particularly in their collapsed state.